+++ org/inbox_test.go
@@ -8,26 +8,60 @@ import (
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestAddInbox(t *testing.T) {
-	key := cbc.Key("test-inbox")
-	st := struct {
-		Inboxes []*org.Inbox
-	}{
-		Inboxes: []*org.Inbox{
+	t.Run("duplicate key", func(t *testing.T) {
+		key := cbc.Key("test-inbox")
+		st := struct {
+			Inboxes []*org.Inbox
+		}{
+			Inboxes: []*org.Inbox{
+				{
+					Key:  key,
+					Code: "BAR",
+				},
+			},
+		}
+		st.Inboxes = org.AddInbox(st.Inboxes, &org.Inbox{
+			Key:  key,
+			Code: "BARDOM",
+		})
+		assert.Len(t, st.Inboxes, 1)
+		assert.Equal(t, "BARDOM", st.Inboxes[0].Code.String())
+	})
+	t.Run("with append nil", func(t *testing.T) {
+		st := struct {
+			Inboxes []*org.Inbox
+		}{
+			Inboxes: []*org.Inbox{
+				{
+					Key:  "test-inbox",
+					Code: "BAR",
+				},
+			},
+		}
+		st.Inboxes = org.AddInbox(st.Inboxes, nil)
+		assert.Len(t, st.Inboxes, 1)
+	})
+	t.Run("with nil list", func(t *testing.T) {
+		inboxes := org.AddInbox(nil, &org.Inbox{Code: "foo"})
+		assert.Len(t, inboxes, 1)
+		assert.Equal(t, "foo", inboxes[0].Code.String())
+	})
+	t.Run("with new inbox", func(t *testing.T) {
+		inboxes := []*org.Inbox{
 			{
-				Key:  key,
+				Key:  "other",
 				Code: "BAR",
 			},
-		},
-	}
-	st.Inboxes = org.AddInbox(st.Inboxes, &org.Inbox{
-		Key:  key,
-		Code: "BARDOM",
+		}
+		inboxes = org.AddInbox(inboxes, &org.Inbox{Key: "test", Code: "FOO"})
+		require.Len(t, inboxes, 2)
+		assert.Equal(t, "BAR", inboxes[0].Code.String())
+		assert.Equal(t, "FOO", inboxes[1].Code.String())
 	})
-	assert.Len(t, st.Inboxes, 1)
-	assert.Equal(t, "BARDOM", st.Inboxes[0].Code.String())
 }
 
 func TestInboxNormalize(t *testing.T) {
@@ -58,6 +92,24 @@ func TestInboxNormalize(t *testing.T) {
 		assert.Equal(t, "BAR", id.Code.String())
 		assert.Equal(t, "0004", id.Ext[iso.ExtKeySchemeID].String())
 	})
+	t.Run("with email in code", func(t *testing.T) {
+		id := &org.Inbox{
+			Code: "dev@invopop.com",
+		}
+		id.Normalize(nil)
+		assert.Empty(t, id.Code.String())
+		assert.Empty(t, id.URL)
+		assert.Equal(t, "dev@invopop.com", id.Email)
+	})
+	t.Run("with url in code", func(t *testing.T) {
+		id := &org.Inbox{
+			Code: "https://inbox.example.com",
+		}
+		id.Normalize(nil)
+		assert.Empty(t, id.Code.String())
+		assert.Empty(t, id.Email)
+		assert.Equal(t, "https://inbox.example.com", id.URL)
+	})
 }
 
 func TestInboxValidate(t *testing.T) {
@@ -84,7 +136,7 @@ func TestInboxValidate(t *testing.T) {
 			Key: "fiscal-code",
 		}
 		err := id.Validate()
-		assert.ErrorContains(t, err, "code: cannot be blank without url")
+		assert.ErrorContains(t, err, "code: cannot be blank without url or email")
 	})
 	t.Run("with URL", func(t *testing.T) {
 		id := &org.Inbox{
@@ -93,12 +145,50 @@ func TestInboxValidate(t *testing.T) {
 		err := id.Validate()
 		assert.NoError(t, err)
 	})
+	t.Run("with invalid URL", func(t *testing.T) {
+		id := &org.Inbox{
+			URL: "https:/inbox",
+		}
+		err := id.Validate()
+		assert.ErrorContains(t, err, "url: must be a valid URL")
+	})
 	t.Run("with code and URL", func(t *testing.T) {
 		id := &org.Inbox{
 			Code: "FOOO",
 			URL:  "https://inbox.example.com",
 		}
 		err := id.Validate()
-		assert.ErrorContains(t, err, "url: mutually exclusive with code")
+		assert.ErrorContains(t, err, "url: must be blank with code or email")
 	})
+	t.Run("with code and email", func(t *testing.T) {
+		id := &org.Inbox{
+			Code:  "FOOO",
+			Email: "dev@invopop.com",
+		}
+		err := id.Validate()
+		assert.ErrorContains(t, err, "email: must be blank with code or url")
+	})
+	t.Run("with email and url", func(t *testing.T) {
+		id := &org.Inbox{
+			Email: "dev@invopop.com",
+			URL:   "https://inbox.example.com",
+		}
+		err := id.Validate()
+		assert.ErrorContains(t, err, "email: must be blank with code or url; url: must be blank with code or email")
+	})
+	t.Run("with email", func(t *testing.T) {
+		id := &org.Inbox{
+			Email: "dev@invopop.com",
+		}
+		err := id.Validate()
+		assert.NoError(t, err)
+	})
+	t.Run("with invalid email", func(t *testing.T) {
+		id := &org.Inbox{
+			Email: "dev@invopop",
+		}
+		err := id.Validate()
+		assert.ErrorContains(t, err, "email: must be a valid email address")
+	})
+
 }